client-pi/requirements.txt
pyzmq>=25.0.0
python-dotenv>=1.0.0
numpy<2.0.0
opencv-python-headless<4.10.0.84
cffi==2.0.0
flatbuffers==25.12.19
mpmath==1.3.0
onnxruntime==1.24.1
packaging==26.0
piper-tts==1.3.0
protobuf==6.33.5
pycparser==3.0
sounddevice==0.5.5
sympy==1.14.0
